Debugging a for loop error.
